Structural project changes
Compiler flags are grouped and saved in variables, this way it is easier
to decide if to use them on all targets or only some.

Since warnings and error flags should not change the behaviour of valid
and good written programs, other flags should have a minor priority, and
not used a priori for all targets.

Flags for enforcing runtime checks (relro, pie, ...) still need to be
sorted out.